+++ CHANGES.md
@@ -2,6 +2,16 @@ Release Notes
 
 _NOTE - This release will work on Moodle 3.9, Moodle 3.10, and Moodle 3.11._
 
+##### 1.39.06 (Build 2021101501)
+New features / improvements:
+
+* Added privacy API.
+* Orphaned ratings are removed.
+
+Big fixes:
+
+* Fixed error with group context.
+
 ##### 1.39.05 (Build 2021092101)
 New features / improvements: